[BPK-4371] Move accordion component to new stories format
